requirements.txt
python-dotenv
tf2-utilities
websockets
uvicorn
aiohttp
fastapi
motor